fix bugs
Add an popup alert window after finishing the first phase)
Restart the numbering at Scroll test 1 for the second phase
Let the grey mask show earlier so that there won't be any time interval between "notice the stars are in the right place" and "the screen greyed out".
Add the summary data to the result.html and in the JSON & CSV file
Change the times shown to just be numbers and not have “ms” on the end and in the CSV and JSON tables
Remove the px in the outputs.
Update the headers of both JSON and CSV files with the units.
Fix the bug of re-input data in the info page. No race condition now.